Refactor of BigBooga
Made a huge refactor of the scripts to allow for scalability
