+++ .idea/workspace.xml
@@ -1,0 +1,91 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="AutoImportSettings">
+    <option name="autoReloadType" value="SELECTIVE" />
+  </component>
+  <component name="ChangeListManager">
+    <list default="true" id="e06f8e6f-21fc-47a9-abca-cc70fee4eaa8" name="Changes" comment="">
+      <change afterPath="$PROJECT_DIR$/.gitignore" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/Agents/Func/initialize_agent.py" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/Agents/Func/initialize_config.py" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/Agents/context_agent.py" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/Agents/execution_agent.py" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/Agents/prioritization_agent.py" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/Agents/task_creation_agent.py" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/Config/config.ini" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/Personas/default.json" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/Personas/load_persona_data.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/LICENSE" beforeDir="false" afterPath="$PROJECT_DIR$/About/LICENSE" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/README.md" beforeDir="false" afterPath="$PROJECT_DIR$/About/README.md" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/config.ini" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/embedding_utils.py" beforeDir="false" afterPath="$PROJECT_DIR$/Utilities/embedding_utils.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/main.py" beforeDir="false" afterPath="$PROJECT_DIR$/main.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/oobabooga_api.py" beforeDir="false" afterPath="$PROJECT_DIR$/APIs/oobabooga_api.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/openai_api.py" beforeDir="false" afterPath="$PROJECT_DIR$/APIs/openai_api.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/pinecone_utils.py" beforeDir="false" afterPath="$PROJECT_DIR$/Utilities/pinecone_utils.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/requirements.txt" beforeDir="false" afterPath="$PROJECT_DIR$/About/requirements.txt" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/task_agents.py" beforeDir="false" />
+    </list>
+    <option name="SHOW_DIALOG" value="false" />
+    <option name="HIGHLIGHT_CONFLICTS" value="true" />
+    <option name="HIGHLIGHT_NON_ACTIVE_CHANGELIST" value="false" />
+    <option name="LAST_RESOLUTION" value="IGNORE" />
+  </component>
+  <component name="FileTemplateManagerImpl">
+    <option name="RECENT_TEMPLATES">
+      <list>
+        <option value="Python Script" />
+      </list>
+    </option>
+  </component>
+  <component name="Git.Settings">
+    <option name="RECENT_GIT_ROOT_PATH" value="$PROJECT_DIR$" />
+  </component>
+  <component name="HighlightingSettingsPerFile">
+    <setting file="file://$PROJECT_DIR$/venv/Lib/site-packages/pinecone/core/client/rest.py" root0="SKIP_INSPECTION" />
+  </component>
+  <component name="MarkdownSettingsMigration">
+    <option name="stateVersion" value="1" />
+  </component>
+  <component name="ProjectId" id="2O9d8Lbo7c6JTYK2XVb4CWzDx1l" />
+  <component name="ProjectLevelVcsManager">
+    <ConfirmationsSetting value="2" id="Add" />
+  </component>
+  <component name="ProjectViewState">
+    <option name="foldersAlwaysOnTop" value="false" />
+    <option name="hideEmptyMiddlePackages" value="true" />
+    <option name="showExcludedFiles" value="false" />
+    <option name="showLibraryContents" value="true" />
+    <option name="showScratchesAndConsoles" value="false" />
+    <option name="sortByType" value="true" />
+  </component>
+  <component name="PropertiesComponent"><![CDATA[{
+  "keyToString": {
+    "RunOnceActivity.ShowReadmeOnStart": "true",
+    "git-widget-placeholder": "main",
+    "ignore.virus.scanning.warn.message": "true",
+    "last_opened_file_path": "F:/Oobabooga/oobabooga-windows",
+    "settings.editor.selected.configurable": "preferences.fileTypes"
+  }
+}]]></component>
+  <component name="RecentsManager">
+    <key name="MoveFile.RECENT_KEYS">
+      <recent name="F:\BigBoogaAGI\Agents\Func" />
+      <recent name="F:\BigBoogaAGI\Agents" />
+      <recent name="F:\BigBoogaAGI\APIs" />
+      <recent name="F:\BigBoogaAGI\About" />
+      <recent name="F:\BigBoogaAGI\Config" />
+    </key>
+  </component>
+  <component name="SpellCheckerSettings" RuntimeDictionaries="0" Folders="0" CustomDictionaries="0" DefaultDictionary="application-level" UseSingleDictionary="true" transferred="true" />
+  <component name="TaskManager">
+    <task active="true" id="Default" summary="Default task">
+      <changelist id="e06f8e6f-21fc-47a9-abca-cc70fee4eaa8" name="Changes" comment="" />
+      <created>1680976570231</created>
+      <option name="number" value="Default" />
+      <option name="presentableId" value="Default" />
+      <updated>1680976570231</updated>
+    </task>
+    <servers />
+  </component>
+</project>
