+++ Agents/Func/initialize_config.py
@@ -1,0 +1,5 @@
+import configparser
+
+config = configparser.ConfigParser()
+config.read('Config/config.ini')
+language_model_api = config.get('LanguageModelAPI', 'library')
