+++ Config/config.ini
@@ -1,0 +1,21 @@
+#Language Model API Options:
+# library = openai_api
+# library = oobabooga_api
+[LanguageModelAPI]
+library = openai_api
+
+[StorageAPI]
+library = pinecone
+
+[EmbeddingLibrary]
+library = sentence_transformers
+
+[OpenAI]
+api_key =
+model_id =
+
+[Pinecone]
+api_key =
+environment = 
+index_name = test-table
+dimension = 768
