+++ Personas/load_persona_data.py
@@ -1,0 +1,7 @@
+import json
+
+
+def load_persona_data(file_path: str) -> dict:
+    with open(file_path, 'r') as json_file:
+        data = json.load(json_file)
+    return data
