--- config.ini
@@ -1,21 +1,0 @@
-#Language Model API Options:
-# library = openai_api
-# library = oobabooga_api
-[LanguageModelAPI]
-library = oobabooga_api
-
-[StorageAPI]
-library = pinecone
-
-[EmbeddingLibrary]
-library = sentence_transformers
-
-[OpenAI]
-api_key = your_openai_api_key
-model_id = your_model_id
-
-[Pinecone]
-api_key = your_pinecone_api_key
-environment = your_pinecone_environment_name
-indexname = your_pinecone_index_name
-dimension = 768
